Merged PR 1792748: Aggregate 5M in Rust
Adding a warmpath aggregate 5M flow for the Rust gateway by following the original design and approach. In summary, we store the information we need in memory and flush every 5 minutes. The data is aggregated based on a key hash of static fields and the values like count, max and min per dimension are upserted.

Related work items: #3761267, #4673400

Signed-off-by: Shuai Tian <shuaitian@microsoft.com>