Merged PR 1827701: [RUM] Split core RUM logic from documentdb_extended_rum to a separate .so file
### Does this PR have any customer impact?
No

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Refactoring

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)

### Are you introducing any new config? If yes, do you have tests with and without them being set?

### ChangeLog (Refer [Template](../oss/CHANGELOG.md))

### Description

documentdb_extended_rum must depend on pg_documentdb to ensure it can register the index extensibility on startup.

The core logic of RUM is unfortunately also embedded in this so file. To support more generic scenarios (in the future as well) where rum and its integration in documentdb are separated, split the core indexing logic out into a separate .so file that doesn't depend on documentdb.

----
#### AI description  (iteration 2)
#### PR Classification
This PR introduces a new feature that integrates a dedicated core module for the Extended RUM index and enhances the build and initialization process.

#### PR Summary
This pull request adds a new core Makefile and updates related build scripts to compile, link, and install the DocumentDB Extended RUM core module while integrating symbol export validation and explicit module initialization.
- `oss/pg_documentdb_extended_rum/core/Makefile`: New file added to define build settings for the core module.
- `oss/pg_documentdb_extended_rum/Makefile`: Updated targets and linking flags to incorporate the core library.
- `pgmongo/Makefile`: Added targets for generating and validating common symbol exports between modules.
- `oss/pg_documentdb_extended_rum/src/documentdb_extended_rum_adapter.c`: Introduced module initialization functions and updated export declarations.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Signed-off-by: Shuai Tian <shuaitian@microsoft.com>