Merged PR 1892820: Fix OSS start server script
### Does this PR have any customer impact?
No

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
BugFix

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No

### Are you introducing any new config? If yes, do you have tests with and without them being set?
No

### ChangeLog (Refer [Template](../oss/CHANGELOG.md))

### Description

Fix OSS start server script, don't check port already in use when we want to stop the server

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix addressing an issue with the server script's port conflict check.

#### PR Summary
This pull request corrects the behavior of the OSS start server script by modifying the port usage check, ensuring that the check is skipped when the server is stopping.
- `/oss/scripts/start_oss_server.sh`: Added a condition to bypass the port check when stopping (i.e., when the stop flag is true), preventing unnecessary errors.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Signed-off-by: Shuai Tian <shuaitian@microsoft.com>